+++ htdocs/app/Http/Controllers/SchoolController.php
@@ -562,7 +562,7 @@ class SchoolController extends Controller
 					$user = $model->newQuery()
 	        		->where('idno', $original['cn'])
 	        		->first();
-	        		$user->delete();
+	        		if ($user) $user->delete();
 					return redirect('school/student?field='.$my_field.'&keywords='.$keywords)->with("success", "已經為您更新學生基本資料！");
 				} else {
 					return redirect('school/student?field='.$my_field.'&keywords='.$keywords)->with("error", "學生身分證字號變更失敗！".$openldap->error());
@@ -1195,7 +1195,7 @@ class SchoolController extends Controller
 	        		$user = $model->newQuery()
 	        		->where('idno', $original['cn'])
 	        		->first();
-	        		$user->delete();
+	        		if ($user) $user->delete();
 					if ($request->user()->idno == $original['cn']) Auth::logout();
 					return redirect('school/teacher?field='.$my_field.'&keywords='.$keywords)->with("success", "已經為您更新教師基本資料！");
 				} else {
